p2p init.

+++ com.jebao.root/1.1-com.jebao.p2p.web.api/src/main/java/com/jebao/p2p/web/api/utils/session/CurrentUser.java
@@ -4,17 +4,18 @@ package com.jebao.p2p.web.api.utils.session;
  * Created by Administrator on 2016/9/22.
  */
 public class CurrentUser {
-    private int Id;
+    private long Id;
     private String Name;
     private String UUID;
 
-    public int getId() {
+    public long getId() {
         return Id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         Id = id;
     }
+
     public String getName() {
         return Name;
     }
